+++ app/services/hybrasyl/constants/body_colors.rb
@@ -1,0 +1,29 @@
+module Hybrasyl
+  module Constants
+    class BodyColors
+      INVISIBLE = 0
+      BLACK = 1
+      RED = 2
+      ORANGE = 3
+      YELLOW = 4
+      TEAL = 5
+      BLUE = 6
+      PURPLE = 7
+      DARK_GREEN = 8
+      GREEN = 9
+      LIGHT_ORANGE = 10
+      BROWN = 11
+      GREY = 12
+      DARK_BLUE = 13
+      FLESH = 14
+      WHITE = 15
+
+      HASH = { "Invisible" => 0, "Black" => 1, "Red" => 2, "Orange" => 3, "Yellow" => 4,
+               "Teal" => 5, "Blue" => 6, "Purple" => 7, "Dark Green" => 8,
+               "Green" => 9, "Light Orange" => 10, "Brown" => 11, "Grey" => 12,
+               "Dark Blue" => 13, "Flesh" => 14, "White" => 15 }
+      REVERSEHASH = HASH.invert
+
+    end
+  end
+end
