+++ app/services/hybrasyl/constants/classes.rb
@@ -1,0 +1,18 @@
+module Hybrasyl
+  module Constants
+    class Classes
+
+      PEASANT = 0
+      WARRIOR = 1
+      ROGUE = 2
+      WIZARD = 3
+      PRIEST = 4
+      MONK = 5
+
+      HASH = {'Peasant' => 0, 'Warrior' => 1, 'Rogue' => 2, 'Wizard' => 3, 'Priest' => 4,
+              'Monk' => 5 }
+      REVERSEHASH = HASH.invert
+
+    end
+  end
+end
