+++ app/services/hybrasyl/constants/elements.rb
@@ -1,0 +1,21 @@
+module Hybrasyl
+  module Constants
+    class Elements
+      NONE = 0
+      FIRE = 1
+      WATER = 2
+      WIND = 3
+      EARTH = 4
+      LIGHT = 5
+      DARK = 6
+      WOOD = 7
+      METAL = 8
+      UNDEAD = 9
+      RANDOM = 10
+      HASH = {'None' => 0, 'Fire' => 1, 'Water' => 2, 'Wind' => 3, 'Earth' => 4,
+              'Light' => 5, 'Dark' => 6, 'Wood' => 7, 'Metal' => 8, 'Undead' => 9, 'Random' => 10 }
+      REVERSEHASH = HASH.invert
+
+    end
+  end
+end
