+++ app/services/hybrasyl/constants/genders.rb
@@ -1,0 +1,13 @@
+module Hybrasyl
+  module Constants
+    class Genders
+      NEUTRAL = 0
+      MALE = 1
+      FEMALE = 2
+
+      HASH = { 'Neutral' => 0, 'Male' => 1, 'Female' => 2 }
+      REVERSEHASH = HASH.invert
+
+    end
+  end
+end
