+++ app/services/hybrasyl/constants/integers.rb
@@ -1,0 +1,19 @@
+module Hybrasyl
+  module Constants
+    class Integers
+      MAX_INT32 = 4294967295
+      MAX_SIGNED32 = 2147483647
+      MIN_SIGNED32 = -2147483647
+      MAX_SIGNED8 = 127
+      MIN_SIGNED8 = -127
+      MAX_INT8 = 255
+      MAX_INT16 = 65535
+      MAX_SIGNED16 = 32767
+      MIN_SIGNED16 = -32767
+
+      MAX_LEVEL = 99
+      MAX_AB = 99
+
+    end
+  end
+end
